+++ src/main/java/co/com/park/gp/business/domain/LoginDomain.java
@@ -6,28 +6,28 @@ import co.com.park.gp.crosscutting.helpers.TextHelper;
 public class LoginDomain {
 
     private TipoEmpleadoDomain tipoEmpleado;
-    private int numeroIdentificacion;
+    private Long numeroIdentificacion;
     private String password;
 
-    public LoginDomain(final TipoEmpleadoDomain tipoEmpleado, final int numeroIdentificacion, final String password) {
+    public LoginDomain(final TipoEmpleadoDomain tipoEmpleado, final Long numeroIdentificacion, final String password) {
         setTipoEmpleado(tipoEmpleado);
         setNumeroIdentificacion(numeroIdentificacion);
         setPassword(password);
     }
 
-    public static LoginDomain build(final TipoEmpleadoDomain tipoEmpleado, final int numeroIdentificacion, final String password) {
+    public static LoginDomain build(final TipoEmpleadoDomain tipoEmpleado, final Long numeroIdentificacion, final String password) {
         return new LoginDomain(tipoEmpleado, numeroIdentificacion, password);
     }
 
     public static LoginDomain build() {
-        return new LoginDomain(TipoEmpleadoDomain.build(), 0, TextHelper.EMPTY);
+        return new LoginDomain(TipoEmpleadoDomain.build(), 0L, TextHelper.EMPTY);
     }
 
     private void setTipoEmpleado(TipoEmpleadoDomain tipoEmpleado) {
         this.tipoEmpleado = ObjectHelper.getObjectHelper().getDefaultValue(tipoEmpleado, TipoEmpleadoDomain.build());
     }
 
-    private void setNumeroIdentificacion(int numeroIdentificacion) {
+    private void setNumeroIdentificacion(Long numeroIdentificacion) {
         this.numeroIdentificacion = numeroIdentificacion;
     }
 
@@ -39,7 +39,7 @@ public class LoginDomain {
         return tipoEmpleado;
     }
 
-    public int getNumeroIdentificacion() {
+    public long getNumeroIdentificacion() {
         return numeroIdentificacion;
     }
